chore: move definitions away from pages, fmt
